Fixed a few warnings.

+++ include/sys/zfs_context_userland.h
@@ -539,7 +539,7 @@ extern void delay(clock_t ticks);
 #define	minclsyspri	60
 #define	maxclsyspri	99
 
-#define	CPU_SEQID	((int)pthread_self() & (max_ncpus - 1))
+#define	CPU_SEQID	((uint64_t)pthread_self() & (max_ncpus - 1))
 
 #define	kcred		NULL
 #define	CRED()		NULL
